+++ README.md
@@ -24,15 +24,8 @@ Qiitaの記事にある「表示しているページのURLをクリップボー
 * Webpack
 * Moment.js
 * jQuery
+* React
 * Jest
-
-## Project Structure
-
-* src/typescript: TypeScript source files
-* src/assets: static files
-* dist: Chrome Extension directory
-* dist/js: Generated JavaScript files
-
 
 ## Setup
