+++ src/Functions/Transition.ts
@@ -16,7 +16,7 @@ export function transitionToNextPage(url: string, newTab: boolean): void {
   chrome.tabs.query({ active: true, currentWindow: true }, (tabs) => {
     const active = tabs[0].id;
     // Set the URL to the Local-NTP (New Tab Page)
-    chrome.tabs.update(active, { url: url }, function () {});
-  })
-  window.close()
+    chrome.tabs.update(active, { url: url });
+  });
+  window.close();
 }
