+++ src/background.ts
@@ -15,7 +15,7 @@ function saveUrlAndTitle(): void {
     alert(
       '現在開いているページのタイトルとURLをクリップボードにコピーしました'
     );
-  })
+  });
 }
 
 function saveUrlAndTitleForMarkDown(): void {
@@ -31,22 +31,7 @@ function saveUrlAndTitleForMarkDown(): void {
     alert(
       '現在開いているページのタイトルとURLをmarkdown形式でクリップボードにコピーしました'
     );
-  })
-}
-
-function unique(array: Array<string>): Array<string> {
-  // JavaScriptのArrayでuniqする8つの方法（と、その中で最速の方法） - Qiita
-  // https://qiita.com/piroor/items/02885998c9f76f45bfa0#object%E3%81%AE%E3%82%AD%E3%83%BC%E3%82%92%E4%BD%BF%E3%81%86%E6%96%B9%E6%B3%95
-  const knownElements = {};
-  const uniquedArray = [];
-  for (let i = 0, maxi = array.length; i < maxi; i++) {
-    if (array[i] in knownElements) {
-      continue;
-    }
-    uniquedArray.push(array[i]);
-    knownElements[array[i]] = true;
-  }
-  return uniquedArray;
+  });
 }
 
 /**
